resumeformatter.onlyoffice/Backend/requirements.txt
# ============================================================================
# RESUME FORMATTER - COMPLETE REQUIREMENTS
# ============================================================================
# This file contains ALL dependencies needed for the Resume Formatter project
# Both for local development and Azure deployment
# ============================================================================

# ============================================================================
# CORE WEB FRAMEWORK
# ============================================================================
Flask==3.0.0                    # Main web framework
Flask-CORS==4.0.0               # Cross-Origin Resource Sharing support
Werkzeug==3.0.1                 # WSGI utility library (comes with Flask)
MarkupSafe==2.1.3               # Required by Flask/Werkzeug
# gunicorn==20.1.0                # Production WSGI server (for Azure deployment) - REMOVED: Duplicate

# ============================================================================
# DOCUMENT PROCESSING - CORE
# ============================================================================
python-docx==1.1.0              # Read/write Word documents (.docx)
pdfplumber==0.10.3              # Extract text and data from PDFs
PyPDF2==3.0.1                   # PDF manipulation and reading
pypdf==3.17.0                   # Modern PDF library
reportlab==4.0.7                # PDF generation
Pillow>=10.3.0                  # Image processing (PIL fork) - Updated for Python 3.13
# PyMuPDF==1.23.8               # Fast PDF processing (fitz) - COMMENTED: Requires Visual Studio on Windows
mammoth==1.6.0                  # Convert DOCX to HTML for preview

# ============================================================================
# MACHINE LEARNING & NLP (Optimized for Speed + Accuracy)
# ============================================================================
# These improve resume parsing accuracy and section detection
# OPTIMIZED: Using lightweight models for 5-10x faster processing!
sentence-transformers>=2.2.2    # Semantic text similarity (uses all-MiniLM-L6-v2)
transformers>=4.35.0            # Hugging Face transformers
torch>=2.1.0                    # PyTorch (required by transformers)
spacy>=3.7.2                    # NLP library for text processing
fuzzywuzzy==0.18.0              # Fuzzy string matching (FAST!)
python-Levenshtein>=0.21.1      # Fast string similarity (for fuzzywuzzy)

# ============================================================================
# AZURE DEPLOYMENT DEPENDENCIES
# ============================================================================
gunicorn==21.2.0                   # WSGI server for production deployment
azure-storage-blob==12.19.0        # Azure Blob Storage for persistent data
python-dotenv==1.0.0               # Load environment variables from .env file

# ============================================================================
# DATA PROCESSING & UTILITIES
# ============================================================================
numpy>=1.26.0                   # Numerical computing (updated for Python 3.13)
# regex==2023.10.3              # Advanced regex operations - COMMENTED: Requires Visual Studio on Windows
python-dateutil==2.8.2          # Date parsing utilities

# ============================================================================
# AZURE MONITORING & ANALYTICS
# ============================================================================
opencensus-ext-azure==1.1.13       # Azure Application Insights integration
opencensus-ext-flask==0.8.2        # Flask middleware for automatic tracking
azure-monitor-opentelemetry==1.2.0 # Modern Azure monitoring SDK

# ============================================================================
# DEVELOPMENT & TESTING (Optional - for local development)
# ============================================================================
# Uncomment these for development
# pytest==7.4.3                 # Testing framework
# pytest-cov==4.1.0             # Code coverage
# black==23.11.0                # Code formatter
# flake8==6.1.0                 # Linting

# ============================================================================
# WINDOWS-SPECIFIC (DO NOT USE ON AZURE)
# ============================================================================
# These packages only work on Windows - DO NOT uncomment for Azure deployment
# docx2pdf                      # Windows-only: Convert DOCX to PDF
# pywin32                       # Windows-only: Windows API access

# ============================================================================
# NOTES FOR DEPLOYMENT
# ============================================================================
# 1. For LOCAL Windows development:
#    - You can install docx2pdf and pywin32 separately if needed
#    - Run: pip install docx2pdf pywin32
#
# 2. For AZURE deployment:
#    - This file works as-is
#    - Azure uses Linux, so Windows packages are excluded
#    - ML packages are optional but recommended for better accuracy
#
# 3. To install ML packages (large download ~2GB):
#    - Run: pip install -r requirements.txt
#    - After install, download spaCy model: python -m spacy download en_core_web_sm
#
# 4. For minimal installation (without ML):
#    - Install only packages above the ML section
#    - The app will work but with reduced accuracy
#
# 5. Installation command:
#    pip install -r requirements.txt
#
# ============================================================================
# TOTAL SIZE ESTIMATE
# ============================================================================
# - Core packages only: ~200MB
# - With ML packages: ~2.5GB
# - Installation time: 5-15 minutes (depending on internet speed)
# ============================================================================
